backend/requirements/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
boto3==1.35.0
django-ses==4.3.1
black==23.11.0
celery==5.3.6
cryptography==42.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.8
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
flake8==6.1.0
gunicorn==21.2.0
iniconfig==2.3.0
kombu==5.6.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
